fix: update function documentation and move to utils
